- added gitignore - moved person.py to top level of sim_objs (the nested file structure is unnecessary) : - moved name files to new top level directory called "resources" - created definitions.py where we will hold constants - made person a child of sim_objs - renamed def age_up to update() in order to override the SimObj::update()
NOTE: I left names.py alone. I did not see where it was used but it
might be important.
