+++ definitions.py
@@ -1,0 +1,5 @@
+import os
+
+ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
+RESOURCES_DIR = os.path.join(ROOT_DIR, "resources")
+
